refactor: extracted `chat` and `step` from cli.py into new chat.py
